common: str_map: get value from map for given key
We add a helper function for a k/v string map so that obtaining a value
from it doesn't become a tedious task.  This function behaviour maps
that of 'get_str_map()'.  As such, we pass a map of string -> string,
the 'key' we're looking for and a 'default value' that's to be returned
in case there is no such key in the map.

The function will return:

- 'default value' if there's no such key.
- 'key' if map['key'] is an empty string
- map['key'] otherwise

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>